Move rendering of dialog icon up to the dialog root
