+++ app/src/ui/dialog/header.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react'
 import { Octicon, OcticonSymbol } from '../octicons'
-import { assertNever } from '../../lib/fatal-error'
 
 interface IDialogHeaderProps {
   /**
@@ -26,14 +25,6 @@ interface IDialogHeaderProps {
    * ways described in the dismissable prop.
    */
   readonly onDismissed?: () => void
-
-  /**
-   * An optional type of dialog header. If the type is error or warning
-   * an applicable icon will be rendered top left in the dialog.
-   *
-   * Defaults to 'normal' if omitted.
-   */
-  readonly type?: 'normal' | 'warning' | 'error'
 
   /**
    * Whether or not the dialog contents are currently involved in processing
@@ -77,29 +68,9 @@ export class DialogHeader extends React.Component<IDialogHeaderProps, {}> {
     )
   }
 
-  private renderIcon() {
-    if (this.props.loading === true) {
-      return <Octicon className="icon spin" symbol={OcticonSymbol.sync} />
-    }
-
-    if (this.props.type === undefined || this.props.type === 'normal') {
-      return null
-    } else if (this.props.type === 'error') {
-      return <Octicon className="icon" symbol={OcticonSymbol.stop} />
-    } else if (this.props.type === 'warning') {
-      return <Octicon className="icon" symbol={OcticonSymbol.alert} />
-    }
-
-    return assertNever(
-      this.props.type,
-      `Unknown dialog header type ${this.props.type}`
-    )
-  }
-
   public render() {
     return (
       <header className="dialog-header">
-        {this.renderIcon()}
         <h1 id={this.props.titleId}>
           {this.props.title}
         </h1>
